fix unity with hmr
